backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
aiohttp==3.11.11
python-dotenv==1.0.1
groq==1.1.2
